Proof of the real interpolation theorem (#115)
Proof of the real interpolation theorem. The proof roughly follows the
exposition in Folland's book, but there are some changes as well, mainly
to get a(n) (nicer) expression for the constant, and because in some
respects it gives a simpler argument. Note also that the proof works for
`NormedAddCommGroups`, and doesn't require the spaces to have a linear
structure, and that only subadditivity, rather than sublinearity, is
assumed for `T`, and finally that the proof works for constants `p0 p1
q1 q0 > 0`.

There is also a way in which the theorem is weaker than the one that was
originally presented in the file, and this is on the aspect of
preserving measurbility of `T`: `T` was previously assumed to preserve
measurability on the union of `Lp0` and `Lp1`, and now it is assumed to
preserver measurability on `Lp`. I do believe that this is unavoidable.

Please note that the order of the arguments in the final constant has
changed. Of course, if this is a problem, I can change them back.

---------

Co-authored-by: Floris van Doorn <fpvdoorn@gmail.com>