Set.Countable.measure_biUnion_le_lintegral and HasWeakType.MB_one (#111)
Partial proofs of `Set.Countable.measure_biUnion_le_lintegral` and
`HasWeakType.MB_one`

- The proof of `Finset.measure_biUnion_le_lintegral` has one `sorry`
left because the version of the Vitali covering lemma that we have uses
closed balls, but we need a statement about open balls. This could be
fixed using a different version of Vitali, or adding an assumption about
the metric space `X`. For example, it would be adequate if we knew that
`interior ∘ closedBall = ball` in `X`.
- The proof of `HasWeakType.MB_one` needs a statement about the
boundedness of the radii in `𝓑`. Of course this is trivial if `𝓑` is
assumed to be finite, but I don't see how to handle it for countable
`𝓑`.