Lemma 7.4.2, `adjointCarleson_adjoint`, `BoundedCompactSupport` (#192)
This proves Lemma 7.4.2 and various prerequisites that will also be
useful for other lemmas.
Main additions are:

In `ForestOperator/AlmostOrthogonality.lean`:
- `adjoint_tree_estimate`: Lemma 7.4.2
- `adjointCarleson_adjoint`: `adjointCarleson` is actually the adjoint
of `carlesonOn`
- `BoundedCompactSupport.adjointCarleson`: `adjointCarleson` has bounded
compact support if the input function does

In `TileStructure.lean`:
- `AEStronglyMeasurable.carlesonOn`: `carlesonOn` is a.e. strongly
measurable
- `BoundedCompactSupport.carlesonOn`: `carlesonOn` has bounded compact
support if the input function does

In `Psi.lean`:
- `exists_bound_of_norm_Ks`: existence of uniform bound for `‖Ks x y‖`
if `x` is in a bounded set

The new file `BoundedCompactSupport.lean` defines a predicate for
functions that are (essentially) bounded, a.e. strongly measurable and
have compact support. It will be convenient to use this as an assumption
on input functions in most lemmas and to use/expand helper lemmas in
this file as needed.

(Edited 12/8/24)