feat: Lemma 5.1.2 (#199)
It was harder than I expected :-)

Highlights:
* In the blueprint, the proof starts with "By Lemma 5.4.8, we can
decompose U_3 as a disjoint union of at most 4n+13 collections...". I
don't know where the 4n+13 comes from (in the end, I got 4n+12). I
didn't find an easy argument here and I had to come up with some
covering lemma relying on the fact that dyadic cubes are disjoint or
included in one another (and it has to play some role, because without
this condition there are counterexamples where you can not extract
disjoint subfamilies just from a bounded stacking condition) -- see
`exists_maximal_disjoint_covering_subfamily` below.
* I refactored `BoundedCompactSupport` (because I needed it for a
measure which was not the ambient measure in the `MeasureSpace`). As
hinted in todos in the files, I went for the version requiring that the
range is bounded, and that the function is strongly measurable
(requiring ae strongly measurable would feel weird here since the other
conditions are everywhere conditions, not almost everywhere conditions)
* In `ProofData`, I added the assumptions that `F` and `G` are bounded
(which is the case in the blueprint)
* I found a better constant than the one of the blueprint (the blueprint
loses `2 ^ {3 a ^ 3}`, but with the very same proof one loses only `2 ^
{a ^ 3}`). I stated the version with the better constant, but also the
one with the weaker constant to match exactly the statement in the
blueprint.
* As usual, I found several lemmas that should be in mathlib. I put them
in the `toMathlib` directory, in files with the same names as the files
where they should be moved, for easier upstreaming.

---------

Co-authored-by: Floris van Doorn <fpvdoorn@gmail.com>