Lemma 7.1.4 (#201)
I added two theorems in the ToMathlib file. We might also consider
generalizing `sum_pow_two_le` (in PointwiseEstimate.lean) and adding it
to Mathlib; Mathlib has some results about geometric sums indexed by
naturals, but none that I could find about geometric sums indexed by
integers.

There were a few (as far as I could tell) mistakes in the blueprint,
which I fixed:

- There were a couple uninteresting typos.
- There was some confusion about the constant `C7_1_4 a`: in the
statement of the theorem, it's 10 * 2 ^ (105a^3), but at the end of the
proof, it's 2 ^ (105a^3). I changed it to 10 * 2 ^ (104a^3).
- I don't think that the step originally described in the blueprint as
"[u]sing the doubling property (1.0.8), the definition of d_p and Lemma
2.1.2" works as written. The doubling property requires x_1 to be in
B_2, which I don't think can be proven in this situation. I used the
monotonicity property (1.0.9) to make the balls big enough that this
condition holds; this requires more applications of (1.0.8) to
compensate, but the extra factors from (1.0.8) eventually get absorbed
into the constant, with the same result as originally in the blueprint.
- The proof claims that (7.1.1) says that 𝒥 is a partition of X, but
(7.1.1) actually says that 𝒥 is a partition of the union of the dyadic
cubes; I don't see any assumption that the dyadic cubes cover all of X.
I changed the reference accordingly. (I noticed that the same claim
about 𝒥 being a partition of X is made in the proof of Lemma 7.2.3, but
I didn't edit that part.)
- Toward the end, the blueprint says "we have s(L) <= s(p) for all p,"
but I think that conclusion is only valid assuming L and p are not
disjoint. That doesn't cause any problems for the proof, because we have
nondisjointness in the relevant case; I edited the blueprint to avoid
the stronger claim.

---------

Co-authored-by: Floris van Doorn <fpvdoorn@gmail.com>