feat: Lemma 5.1.3 (#200)
The proof diverges just a little bit from the blueprint because some
already formalized lemmas don't match exactly the blueprint statement.
For instance, Lemma 5.5.2 in the blueprint involves n sets, while the
Lean version uses n+1 sets (which is slightly weaker, but enough to
complete the proof of Lemma 5.1.3). Both versions are true (it would
even be true with n / 50 sets). I don't know if you want to reconcile
the blueprint and the formalization, by changing one or the other -- or
just let it go, and accept local divergences like that which have no
consequence whatsoever.

---------

Co-authored-by: Floris van Doorn <fpvdoorn@gmail.com>