feat: complete proof of Lemma 8.0.1 (#278)
I've also made a few changes throughout the files to get the right
statement of Lemma 8.0.1:
* instead of the norms `iLipNorm` (Lipschitz) and `hnorm` (Hölder), we
now have versions defined using sups in ENNReal, called `iLipENorm` and
`iHolENorm`, and their images in NNReal called `iLipNNNorm` and
`iHolNNNorm`.
* The definition of the Hölder norm has been fixed. It used to be
```lean
⨆ (x ∈ ball x₀ R), (‖ϕ x‖₊ : ℝ≥0∞) +
  R ^ τ * ⨆ (x ∈ ball x₀ R) (y ∈ ball x₀ R) (_ : x ≠ y), (‖ϕ x - ϕ y‖₊ / (nndist x y) ^ τ : ℝ≥0∞)
```
where the second sup was included in the first sup because of the lack
of parentheses.
* The definition of the class `IsCancellative` has been fixed: it
required the test function to be globally Lipschitz, while the right
condition is Lipschitz on a ball
* With the change of `IsCancellative`, the proof that it is satisfied on
the real line becomes harder. I had to change a little bit the Lean
statement and proof of the classical Van der Corput lemma (requiring
only a Lipschitz function on a ball) -- now it matches the statement in
the blueprint.
* I have streamlined the proof of `IsCancellative` on the real line,
using new API around the Lipschitz and Hölder norms
* For doubling measures, I have added the statement
```lean
lemma measure_ball_le_same'' (x : X) {r t : ℝ} (ht : 0 < t) (h't : t ≤ 1) :
    μ.real (ball x r) ≤ A * t ^ (- Real.logb 2 A) * μ.real (ball x (t * r))
```
Once this is established, this avoids juggling with powers of 2 in the
proof of Lemma 8.0.1. I don't know if there are other places in the
files where it could be used to streamline arguments.
* Use enorm instead of the coercion of nnnorm wherever I've noticed it

I can split in in smaller PRs if you prefer.

---------

Co-authored-by: Floris van Doorn <fpvdoorn@gmail.com>