chore: bump mathlib to 4.19.0rc2 (#293)
Builds on #287 (but I couldn't push to this branch because I work from a
fork, hence the new PR).

The issue that #287 couldn't solve is solved here in an ugly way
(declaring an instance by hand with `letI`). Probably related to the new
structure elaboration. I'll report the difficulty on Zulip.

The file `MeasureReal` has been upstreamed, and is therefore deleted in
this PR. During the upstreaming, there was the comment that the `nnreal`
version had no API, and was probably not a good idea anyway, so it has
been deleted. This means that, in the project, I have replaced the lines
`volume.nnreal s` with `(volume s).toNNReal`. In most cases, these are
an indication that one should probably work with ENNReal instead (just
like we switched from NNNorm to ENorm), but I haven't made any such
change in the PR, which is a pure bump PR without any other change.

---------

Co-authored-by: Pietro Monticone <38562595+pitmonticone@users.noreply.github.com>
Co-authored-by: Michael Rothgang <rothgang@math.uni-bonn.de>