feat(WeakType): generalise most remaining lemmas to enorm (#300)
One section about the layercake formula remains (and will be changed in
a future PR). This requires generalising a handful of mathlib lemmas:
https://github.com/leanprover-community/mathlib4/pull/23707 contained
some of these, others need ENormedSpace to be in mathlib first.

For most of this file, there is a trade-off about generality to be made:
currently, there is no typeclass unifying both a normed space over a
field and an ENormedSpace. Hence, we usually add enorm variants of
existing lemmas and keep the existing versions.
We make one change, though: the normed space becomes primed; the enorm
version stays unprimed.

---------

Co-authored-by: Floris van Doorn <fpvdoorn@gmail.com>