(Some steps towards) Task 118 (#294)
This is a proof of `hasStrongType_maximalFunction_todo`, which is task
118 or is part of task 118.

The proof fully compiles but probably should still be edited quite
heavily: In particular, currently a large part of the proof is to work
around the fact that `hasStrongType_maximalFunction` (still) contains
the |>.toReal applied to the maximal function. I think eventually, this
will be removed, and then the proof can be shorted considerably. I am
currently including the code to make it work, because elements of it may
be useful in other places.

---------

Co-authored-by: Floris van Doorn <fpvdoorn@gmail.com>