feat: run environment linters (#309)
Adds two new steps to CI, ensuring that
- Carleson.lean is up to date
- running `runLinter` to check for no warnings
I'm not sure if this actually errors if there are linter warnings (and
am happy to tweak this in the future to ensure this).

This PR also adds a nolint file with the current exceptions, mostly
around the docBlame linter outside of the `ToMathlib` directory.