WIP: use enorm in BoundedCompactSupport (#311)
* Generalize the functions for which we can write
`BoundedCompactSupport`, including functions into `ENNReal`.
* Change the properties so that the function is only almost everywhere
bounded and only a.e.-measurable.
* This leads to a bunch of backwards incompatibilities. In particular we
now have to use `AE(Strongly)Measurable` instead of
`(Strongly)Measurable`
* There are still two sorry's left, where we take a supremum over all
values of a function with `BoundedCompactSupport`. Either these proofs
need to be rewritten to use `essSup` or we have to prove the (true)
stronger claim that `adjointCarleson` is bounded if it's input function
is (only) a.e.-bounded.