Lower-semicontinuity of maximal function - Task 135 (#351)
Proof of the lower-semicontinuity of the maximal function and
lower-semicontinuity of the global maximal function.

Although not used in the proof, I also proved `continuous_average_ball`
under an extra assumption. Without that assumption I proved
`lowerSemicontinuous_integral_ball`. I'm not sure these are still
useful, because they are currently not used. Feel free to discard.