Cleanup of Lemma 6.1.5 proof + natural powers (#380)
The first commit fills in the last sorry in Lemma 6.1.5's proof. To
facilitate the cleanup, in the second commit I changed some exponents
appearing in constants to be `ℕ`-valued rather than `ℝ`- or
`ℝ≥0`-valued, where such a change was possible:
```lean
/-- The constant appearing in part 3 of Lemma 2.1.3. -/
def D2_1_3 (a : ℕ) : ℝ≥0 := 2 ^ (150 * a ^ 3)
```
Overall this leads to fewer extra assumptions that need to be proven and
fewer type conversion steps.