Fix definition of GeneralCase (partial progress on Lemma 10.2.5) (#386)
- The definition of `GeneralCase` had the wrong inequality. This is now
fixed.
- The following two lemmas about `GeneralCase` are proven using the
correct definition.
- The assumption `(hf : BoundedFiniteSupport f)` turned out to be
unnecessary for the proof of `isOpen_MB_preimage_Ioi`, and thus
unnecessary for a lot of subsequent work. I removed the assumption from
any definitions/lemmas that were already completed without using it, but
left it as an assumption for any lemma that doesn't have a proof yet.