Lemma 10.1.8 (#396)
It turned out to be easier to first do 10.1.8 to know the exact
statements for 10.1.7 needed for the continuity argument to follow
through (I used `ContinuousWithinAt.closure_le`).
This will hopefully help to formalize 10.1.7 in a nicer way than to
follow the epsilon-delta argument in the blueprint directly.

I didn't add the `leanok` yet because the statements of the
`small_annulus` lemmata will still change and that will slightly affect
the proof.

---------

Co-authored-by: Michael Rothgang <rothgang@math.uni-bonn.de>