Lemma 10.1.7 (#398)
With the correct statement in place in 10.1.8, it was feasible to prove
both estimates without resorting to epsilon-delta intricacies.
Instead `continuousWithinAt_of_dominated` was used.

This led to intricacies because continuity of an indicator had to be
proved. This was feasible but only at the boundary. Therefore no
`ContinuousOn` or `Continuous` statement was shown.
In fact if you look precisely the formalised statement aligns with the
blueprint in the sense that they show the same thing.

In terms of the formalisation, I struggled to play nicely with
`nhdsWithin` on the reals, so it might be that some statements can be
optimised. Also the proofs look similar, but in critical parts the
left-or-rightness is being used so I'm afraid it cannot be easily
unified.