Proof of Lemma 10.2.5 (#395)
- I had to weaken the conclusion of `tsum_volume_czBall3_le` because
`maximal_theorem` (Lemma 10.2.1) is formalized with a weaker constant
than appears in the blueprint. I changed the statement and proof of
Lemma 10.2.5 in the blueprint, but I didn't change Lemma 10.2.1 because
I wasn't prepared to fix its proof.
- I had to strengthen the assumptions in `eLpNorm_czRemainder_le` and
`tsum_eLpNorm_czRemainder_le`, but that's because the original
formalization was missing an assumption present in the blueprint, so I
don't think that should cause any problems.