Proof of Lemma 10.2.6 (#401)
Proof of Lemma 10.2.6 (`estimate_good`). Includes material from the
proof of Lemma 10.2.5 (#395).

- I fixed the conclusion of the lemma, which was incorrectly formalized
using `czApproximation f ha α` rather than `czApproximation f ha
(c10_0_3 a * α)` (corresponding to the fact that `α'` plays the role of
`α` in the blueprint).
- The formalization of the lemma didn't include the assumption `hT` from
Lemma 10.0.3, so I added it. As far as I know, Lemma 10.2.6 is only used
to prove Lemma 10.0.3, so I don't think there's any harm in that.
- In order to apply `hT`, we needed to know that the `czApproximation`
of a function of bounded finite support is also of bounded finite
support. This was previously not stated in Lemma 10.2.5, so I added the
extra statement and proof to the blueprint.

---------

Co-authored-by: grunweg <rothgang@math.uni-bonn.de>