Lemma 10.0.2 (#417)
The result has been shown up to a variant of `lintegral_iSup'`, namely
`eLpNorm_iSup'`.
The exact statement and formalization of that result is something I have
put off for now, but at least it should hold for `p = 2` which is how it
is used here.

It includes the formalization of measurability, analogous to
`simpleNontangentialOperator`.

The constant had to be adjusted to fit the 10.1.6 constant (polynomial
in `a`) into the 10.0.2 constant (monomial `a^3`). This requires
downstream adjustments in Thm 10.0.1 which I will take along in the PR
for that theorem, including the blueprint changes associated.