Proof of Lemma 10.2.7 (#425)
One TODO left in a comment to figure out where a new ToMathlib lemma
should go, but otherwise this is finished. I don't think the lemma has
an obvious home, so suggestions are welcome.

I made the blueprint changes to get rid of the B_j* / B_j change in
notation (and fix a couple unrelated typos).