change the definition of linearizedNontangentialOperator (#444)
reformulate the definition of nontangentialOperator

The operator `linearizedNontangentialOperator` has been changed,
although the "has bounded weak-type"-condition is equivalent to the
previous one. Lars made these changes in the blueprint today to solve an
issue in the proof of Theorem 1.0.2

Another option was to not change the definition, but one motivating idea
of the blueprint is that integrals we deal with are over a integration
domain with finite measure, and if we don't change this definition,
`linearizedNontangentialOperator` would be an exception to that.