feat: Lemma 11.1.6 (#452)
I had to modify a little bit the blueprint, as it claimed an inequality
was always true, but I could only prove it in some range. If one writes
down the details of the proof, only this range is needed, so the global
strategy of proof works perfectly well modulo these minor adjustments.

I made the following Lean change: the `L^2` norm of `f` on an interval
`[a, b]` was used as `eLpNorm (indicator (Icc a b) f) 2` (where there is
a measure there, the volume, which is implicit). I changed it to
`eLpNorm f 2 (volume.restrict (Icc a b))` here and there, as this is how
we do integrals on sets and it is much more convenient to use since we
have a big API for restrictions of measures (and avoiding indicators is
always a good idea as they are pretty painful).