feat: optimize the constants (#458)
Instead of letting `D = 2 ^ {100 * a^2}`, let `D = 2 ^ {c * a^2}` for
some fixed natural `c`, and express all the constants in terms of `c`
(still keeping all the relevant constants of the form `2 ^ {n * a^2}`
for `n = n(c)`, because it wouldn't make sense to overoptimize). For `c
= 100`, we should recover essentially the constants in the blueprint.
However, we can take `c = 7`, which gives better constants in the end.

---------

Co-authored-by: grunweg <rothgang@math.uni-bonn.de>