Cleanup TwoSidedCarleson / MetricCarleson folders (#469)
The TwoSidedCarleson folder still contained a number of general results that were supposed to go in ToMathlib.
This PR addresses that and other minor fixes.
The MetricCarleson folder was also checked with only minor changes.

I was not sure where to put `iSup_rpow` so I put it in Misc.