Cleanup base folder (except `Psi.lean`) (#476)
More cleanup, this time the base folder. Main focus was to ensure all
ToMathlib elements were caught, and to fix (whitespace) style of the
statements. In some files the proofs can be further optimized as well
but IMO that does not have the priority at this moment.

The reason to omit `Psi.lean` is because it contains a lot of stuff
formalized using `norm` instead of `enorm` and it might be better to
refactor into `enorm` entirely (as the `enorm` versions are used
downstream).
And that seemed like a task better put in a separate PR.