WIP: Proving the HasStrongType upgrade of the main theorem (#498)
This PR works towards proving the upgrade
`two_sided_metric_carleson_strong_type` of the main theorem, the key
result to derive the L^p version of the classical Carleson-Hunt theorem.
The chosen approach (see the file `RestrictedWeakType.lean`) uses
Lorentz space interpolation. Proving this is a separate project but to
be able to apply it here, the main theorem first needs to be upgraded to
`hasLorentzType` via `HasRestrictedWeakType.hasLorentzType` (which
relies on intermediate results in `Lorentz.lean`).

Not all proofs are sorry-free but the overall structure has started to
converge and hopefully, not too many assumptions are missing anymore.

---------

Co-authored-by: Floris van Doorn <fpvdoorn@gmail.com>