Cleanup `ProofData`, extract `iHolENorm` and `iLipENorm` (includes task 401) (#493)
While cleaning up `ProofData.lean`, I decided to expand a bit and move
the definitions of `iHolENorm` and `iLipENorm` to separate files.
Biggest change is to make the tau parameter explicit, IMO this could not
be avoided.
There were some unexpected consequences of that because tau is an abbrev
and so is unfolded by any `simp` call.

I wasn't sure what ToMathlib folder these should go in, so I put them
under the main folder. I played around a bit with the assumptions to try
and minimize them.
Probably (quite) some work is required to bring these to Mathlib-worthy
completeness, generality and quality, it is expressly not my aim to
follow up on that (lacking the specific knowledge to determine
appropriate generality etc.).

But at least now the definitions and the results are completely
abstracted from the rest of Carleson.